Feat: added countdown and a way to initialize a question
The countdown textbox has a separate ID so it is easily accessible. I also created the MultiplayerQuestionCtrl class. In Main.java, there is now a showQuestion() method, that activates this scene. Bug: for some reason, the images aren't showing when I'm running the client, maybe someone has encountered this problem and can help. Thanks!