Fix: changed layout of scene and ID's of elements
I changed the layout in such a way that the main container is now a VBox instead of a HBox and I set the Min Width, Min Height, Pref Width and Pref Height to the appropriate values.
Furthermore, I deleted the panes for correct players and the slider next to it and replaced it with a ListView.