Feat: answering redirects to answer page
Added a general method for handling incoming answers, and one proxy method each for the buttons, as we cannot pass parameters through FXML injection. 
Note that in the final game, this method shouldn't actually redirect to the answer page, just save the game state and maybe calculate points. The redirection will take place when the timer runs out.