Feat: Add timer question page and automatically redirects to answer page
Adds a numerical timer to the question page that counts down from 10 to 0. When the timer reaches zero, the answer buttons are disabled and the player is redirected to the answer page. Moreover, the time it took to click on an answer is stored in a field every time an answer is clicked.