Added the ranking page
Added the ranking page layout with a timer that goes down from 10 and transitions to the question screen when the timer hits zero. Since we don't keep track of question number yet (at least within this branch) I didn't add the functionality to redirect to the ranking page when we are midway. However, there is a showRanking method within MainCtrl.