Feat: Add user starting solo game to repository
Now a user that starts a solo player game will be stored in the user repository. For this, I made a separate repository, but for some reason, the users will be stored in the same database as the multiplayer users. For this reason I added a boolean that identifies solo players. Furthermore, I added a game ID for the users for later use if we want to have multiple multiplayer games at the same time. I changed the logic that checks if a username is already used, since only in multiplayer games usernames must be unique. This means that the boolean that identifies solo players is now also checked.