Feat/Fix: Add time factor to point calculation of estimation question
To encourage players to answer the estimation questions as quickly as possible, I added the time spent by the user on the question as a factor in the calculation of the points the user got for the question. This is already implemented for the other question types, but now the estimation question also has this feature.