Feat: Add frame of multiplayer rematch functionality
I added a method that sets a boolean to the other value when the rematch button is clicked. When the timer ends, the user is redirected to the home page when the boolean is false and to the waiting page if the boolean is true. The user isn't unbinded, since I want to keep the same users for the rematch. I still need to add a way so that the players that want to rematch are put in a separate waiting room. I want to achieve this using the game ID and by changing the query to use a parameter to only fetch users with a corresponding game ID (null for the normal waiting room).