Feat: Add working rematch button to multiplayer results page
I made the rematch button working in on the multiplayer results page. This means that now, when a player clicks on rematch, the player will be automatically redirected to the first question of the new game again after the timer runs out. The game object will be the same, but only the players who will restart will be left and the game will have new questions.