Fix: Begin the emoji path substring from index 0 for inclusivity
Now everyone, no matter the path the user uses for storing the application, can use the game and the emojis without getting an error. Before this, the "file/:" part in the front was omitted, but this caused errors for user with spaces in their paths, now this isn't the case anymore.