Fix: Enable jokers in rematch and enable rematch button after quit
When jokers where used in a match and after that the user had a rematch, the jokers where disabled, but not anymore now. Also, when a user waited too long on the multiplayer results page and the rematch button was disabled, the rematch button was still disabled in the next game, so that is fixed now as well.